+++ pelutils/tests.py
@@ -1,14 +1,20 @@
+from __future__ import annotations
+
 import multiprocessing as mp
 import os
 import sys
+import tempfile
+from functools import wraps
 from shutil import rmtree
+from typing import Callable, TypeVar
 
+_C = TypeVar("_C", bound=Callable)
 
-def restore_argv(fun):
-    """
-    Decorator function that restores sys.argv after function exits
-    This is useful for testing command line argument handling
-    ```
+def restore_argv(fun: _C) -> _C:
+    """Decorator function to restore sys.argv after function exit.
+
+    This is useful for testing command line argument handling.
+    ```py
     @restore_argv
     def test_my_function():
         sys.argv = ["test", "with", "different", "value", "of", "sys.argv"]
@@ -17,7 +23,8 @@ def restore_argv(fun):
     test_my_function()
     # sys.argv still has value x here
     ```
-    """
+    """  # noqa: D401
+    @wraps(fun)
     def wrapper(*args, **kwargs):
         old_argv = sys.argv.copy()
         try:
@@ -27,13 +34,14 @@ def restore_argv(fun):
     return wrapper
 
 class SimplePool:
-    """pytest-cov does not exit properly when using mp.Pool. Using
-    this class as a basic drop-in replacement solves it. For details,
-    see https://github.com/pytest-dev/pytest-cov/issues/250.
+    """Drop-in replacement for mp.Pool for when using it within a pytest context.
+
+    pytest-cov does not exit properly when using mp.Pool. Using this class instead solves it.
+    For details, see https://github.com/pytest-dev/pytest-cov/issues/250.
     """
 
-    def __init__(self, processes=mp.cpu_count()):
-        self._processes = processes
+    def __init__(self, processes: int | None = None):
+        self._processes = processes or mp.cpu_count()
         self._pool = None
 
     def __enter__(self):
@@ -46,7 +54,9 @@ class SimplePool:
         self._pool = None
 
 class UnitTestCollection:
-    """A convenience class for inheriting from when writing test classes using pytest.
+    """
+    A convenience class for inheriting from when writing test classes using pytest.
+
     This class ensures that test directory is automatically created and deleted between tests.
 
     See this example for usage:
@@ -60,20 +70,20 @@ class UnitTestCollection:
     ```
     """
 
-    # Place temporary test files here
-    # Directory will be creating when running test files and removed afterwards
-    test_dir = ".local_test_files"
+    test_dir = tempfile.mkdtemp()
 
     @classmethod
     def setup_class(cls):
+        """Create temporary directory for putting test files."""
         os.makedirs(cls.test_dir, exist_ok=True)
 
     @classmethod
     def teardown_class(cls):
+        """Clean up temporary directory."""
         rmtree(cls.test_dir, onerror=cls.ignore_absentee)
 
     @staticmethod
-    def ignore_absentee(_, __, exc_inf):
+    def ignore_absentee(_, __, exc_inf):  # noqa: D102
         except_instance = exc_inf[1]
         if isinstance(except_instance, FileNotFoundError):
             return
@@ -81,4 +91,8 @@ class UnitTestCollection:
 
     @classmethod
     def test_path(cls, path: str) -> str:
+        """Return a path inside the test directory.
+
+        `path` would often just be a filename which can be written to and is automatically cleaned up after the test.
+        """
         return os.path.join(cls.test_dir, path)